+++ test/add_dependency/app/main.cpp
@@ -5,13 +5,21 @@
 
 #include <iostream>
 
-#include <foo/foo.h>
 #include <baz/baz.h>
+#include <foo/foo.h>
+
+#ifdef EXTRA
+#include <extra/extra.h>
+#endif
 
 int main()
 {
     std::cout << "app" << std::endl;
     std::cout << "foo::whoooth" << foo::whoooth() << std::endl;
     std::cout << "baz::whoooth" << baz::whoooth() << std::endl;
+
+#ifdef EXTRA
+    std::cout << "extra::cowabunga" << extra::cowabunga() << std::endl;
+#endif
     return 0;
 }