+++ test/add_dependency/libextra/src/extra/extra.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include <string>
+
+namespace extra
+{
+std::string cowabunga()
+{
+    return "cowabunga";
+}
+}