+++ test/add_dependency/test_add_dependency.py
@@ -3,6 +3,9 @@
 
 import os
 import json
+import pytest
+from pytest_testdirectory.runresulterror import RunResultError
+
 
 """ Integration testing of adding a dependency.
 
@@ -11,11 +14,11 @@ This test is a bit involved so lets try to explain what it does:
 We are setting up the following dependency graph:
 
            +--------------+
-           |     app      |
-           +---+------+---+
-               |      |
-               |      |
-      +--------+      +-------+
+           |     app      |----optional----+
+           +---+------+---+                |
+               |      |               +----+-----+
+               |      |               | libextra |
+      +--------+      +-------+       +----+-----+
       |                       |
       v                       v
 +-----+------+          +-----+-----+   submodule  +--------+
@@ -29,7 +32,7 @@ We are setting up the following dependency graph:
 
 The arrows indicate dependencies, so:
 
-- 'app' depends on 'libfoo' and 'libbaz'(internal & optional)
+- 'app' depends on 'libfoo' and 'libbaz'(internal)
 - 'libfoo' depends on 'libbar'
 - 'libbar' depends on 'libbaz'
 
@@ -40,18 +43,6 @@ def mkdir_app(directory):
     app_dir = directory.mkdir("app")
     app_dir.copy_file("test/add_dependency/app/main.cpp")
     app_dir.copy_file("test/add_dependency/app/wscript")
-
-    app_dir.copy_file("test/fake_git.py")
-
-    app_dir.copy_file("build/waf")
-
-    return app_dir
-
-
-def mkdir_app_json(directory):
-    app_dir = directory.mkdir("app")
-    app_dir.copy_file("test/add_dependency/app/main.cpp")
-    app_dir.copy_file("test/add_dependency/app/wscript_json", rename_as="wscript")
     app_dir.copy_file("test/add_dependency/app/resolve.json")
 
     app_dir.copy_file("test/fake_git.py")
@@ -75,61 +66,9 @@ def mkdir_app_json(directory):
     return app_dir
 
 
-def mkdir_app_override(directory):
-    app_dir = directory.mkdir("app")
-    app_dir.copy_file("test/add_dependency/app/main.cpp")
-    app_dir.copy_file("test/add_dependency/app/wscript_json", rename_as="wscript")
-    app_dir.copy_file(
-        "test/add_dependency/app/resolve_override.json", rename_as="resolve.json"
-    )
-
-    app_dir.copy_file("test/fake_git.py")
-    app_dir.copy_file("build/waf")
-    # Note: waf will call "git config --get remote.origin.url" in this folder,
-    # so "git init" is required if the pytest temp folder is located within
-    # the main waf folder
-
-    git_info = {
-        "checkout": "master",
-        "branches": ["master"],
-        "commits": [],
-        "remote_origin_url": "git@github.com/acme-corp/app.git",
-        "is_detached_head": False,
-        "submodules": [],
-        "tags": [],
-    }
-
-    json_path = os.path.join(app_dir.path(), "git_info.json")
-
-    with open(json_path, "w") as json_file:
-        json.dump(git_info, json_file)
-
-    return app_dir
-
-
-def commit_file(directory, filename, content):
-    directory.write_text(filename, content, encoding="utf-8")
-    directory.run(["git", "add", "."])
-    directory.run(
-        [
-            "git",
-            "-c",
-            "user.name=John",
-            "-c",
-            "user.email=doe@email.org",
-            "commit",
-            "-m",
-            "oki",
-        ]
-    )
-
-
 def mkdir_libfoo(directory):
     # Add foo dir
-    foo_dir = directory.mkdir("libfoo")
-    foo_dir.copy_file("test/add_dependency/libfoo/wscript")
-    foo_dir.copy_file("test/add_dependency/libfoo/some_repos_contain")
-    foo_dir.copy_dir(directory="test/add_dependency/libfoo/src")
+    foo_dir = directory.copy_dir(directory="test/add_dependency/libfoo")
 
     git_info = {
         "checkout": "master",
@@ -138,7 +77,7 @@ def mkdir_libfoo(directory):
         "remote_origin_url": "git@github.com/acme-corp/foo.git",
         "is_detached_head": False,
         "submodules": [],
-        "tags": ["1.3.3.7"],
+        "tags": ["v1.2.3", "2.3.3.7", "1.3.3.7"],
     }
 
     json_path = os.path.join(foo_dir.path(), "git_info.json")
@@ -146,38 +85,31 @@ def mkdir_libfoo(directory):
     with open(json_path, "w") as json_file:
         json.dump(git_info, json_file)
 
-    foo_dir.write_text(filename="ok.txt", data="hello world", encoding="utf-8")
-
     return foo_dir
 
 
-def mkdir_libfoo_json(directory):
-    # Add foo dir
-    foo_dir = directory.mkdir("libfoo")
-    foo_dir.copy_file("test/add_dependency/libfoo/wscript_json", rename_as="wscript")
-    foo_dir.copy_file("test/add_dependency/libfoo/some_repos_contain")
-    foo_dir.copy_file("test/add_dependency/libfoo/resolve.json")
-    foo_dir.copy_dir(directory="test/add_dependency/libfoo/src")
+def mkdir_libextra(directory):
+    # Add extra dir
+    extra_dir = directory.mkdir("libextra")
+    extra_dir.copy_file("test/add_dependency/libextra/wscript")
+    extra_dir.copy_dir(directory="test/add_dependency/libextra/src")
 
     git_info = {
         "checkout": "master",
         "branches": ["master"],
         "commits": [],
-        "remote_origin_url": "git@github.com/acme-corp/foo.git",
+        "remote_origin_url": "git@github.com/acme-corp/extra.git",
         "is_detached_head": False,
         "submodules": [],
-        "tags": ["1.3.3.7"],
+        "tags": ["2.0.1", "2.0.0"],
     }
 
-    json_path = os.path.join(foo_dir.path(), "git_info.json")
+    json_path = os.path.join(extra_dir.path(), "git_info.json")
 
     with open(json_path, "w") as json_file:
         json.dump(git_info, json_file)
 
-    foo_dir.write_text(filename="ok.txt", data="hello world", encoding="utf-8")
-    # commit_file(directory=foo_dir, filename="ok.txt", content=u"hello world")
-
-    return foo_dir
+    return extra_dir
 
 
 def mkdir_libbar(directory):
@@ -414,42 +346,7 @@ def run_commands(app_dir, git_dir):
     app_dir.run(["python", "waf", "build", "-v"])
 
 
-def test_resolve_json(testdirectory):
-    """Tests that dependencies declared in the wscript works. I.e. where we
-    call add_dependency(...) in the resolve function of the wscript.
-    """
-
-    app_dir = mkdir_app_json(directory=testdirectory)
-
-    # Make a directory where we place the libraries that we would have cloned
-    # if we had use the full waf resolve functionality.
-    #
-    # To avoid relying on network connectivity we simply place the
-    # libraries there and then fake the git clone step.
-    git_dir = testdirectory.mkdir(directory="git_dir")
-
-    qux_dir = mkdir_libqux(directory=git_dir)
-    foo_dir = mkdir_libfoo_json(directory=git_dir)
-    bar_dir = mkdir_libbar(directory=git_dir)
-    baz_dir = mkdir_libbaz(directory=git_dir, qux_dir=qux_dir)
-
-    # Instead of doing an actual Git clone - we fake it and use the paths in
-    # this mapping
-    clone_path = {
-        "acme-corp/foo.git": foo_dir.path(),
-        "acme-corp/bar.git": bar_dir.path(),
-        "acme/baz.git": baz_dir.path(),
-    }
-
-    json_path = os.path.join(app_dir.path(), "clone_path.json")
-
-    with open(json_path, "w") as json_file:
-        json.dump(clone_path, json_file)
-
-    run_commands(app_dir=app_dir, git_dir=git_dir)
-
-
-def test_add_dependency(testdirectory):
+def test_resolve(testdirectory):
     """Tests that dependencies declared in the wscript works. I.e. where we
     call add_dependency(...) in the resolve function of the wscript.
     """
@@ -467,6 +364,7 @@ def test_add_dependency(testdirectory):
     foo_dir = mkdir_libfoo(directory=git_dir)
     bar_dir = mkdir_libbar(directory=git_dir)
     baz_dir = mkdir_libbaz(directory=git_dir, qux_dir=qux_dir)
+    extra_dir = mkdir_libextra(directory=git_dir)
 
     # Instead of doing an actual Git clone - we fake it and use the paths in
     # this mapping
@@ -474,6 +372,7 @@ def test_add_dependency(testdirectory):
         "acme-corp/foo.git": foo_dir.path(),
         "acme-corp/bar.git": bar_dir.path(),
         "acme/baz.git": baz_dir.path(),
+        "acme-corp/extra.git": extra_dir.path(),
     }
 
     json_path = os.path.join(app_dir.path(), "clone_path.json")
@@ -492,12 +391,14 @@ def test_add_dependency_path(testdirectory):
     qux_dir = mkdir_libqux(directory=git_dir)
     foo_dir = mkdir_libfoo(directory=git_dir)
     bar_dir = mkdir_libbar(directory=git_dir)
+    extra_dir = mkdir_libextra(directory=git_dir)
 
     # Instead of doing an actual Git clone - we fake it and use the paths in
     # this mapping
     clone_path = {
         "acme-corp/foo.git": foo_dir.path(),
         "acme-corp/bar.git": bar_dir.path(),
+        "acme-corp/extra.git": extra_dir.path(),
     }
 
     json_path = os.path.join(app_dir.path(), "clone_path.json")
@@ -540,6 +441,7 @@ def test_create_standalone_archive(testdirectory):
     foo_dir = mkdir_libfoo(directory=git_dir)
     bar_dir = mkdir_libbar(directory=git_dir)
     baz_dir = mkdir_libbaz(directory=git_dir, qux_dir=qux_dir)
+    extra_dir = mkdir_libextra(directory=git_dir)
 
     # Instead of doing an actual Git clone - we fake it and use the paths in
     # this mapping
@@ -547,6 +449,7 @@ def test_create_standalone_archive(testdirectory):
         "acme-corp/foo.git": foo_dir.path(),
         "acme-corp/bar.git": bar_dir.path(),
         "acme/baz.git": baz_dir.path(),
+        "acme-corp/extra.git": extra_dir.path(),
     }
 
     json_path = os.path.join(app_dir.path(), "clone_path.json")
@@ -569,48 +472,6 @@ def test_create_standalone_archive(testdirectory):
     assert app_dir.contains_file("test_add_dependency-1.0.0.zip")
 
 
-def test_override_json(testdirectory):
-    """Tests that dependencies marked as override works"""
-
-    app_dir = mkdir_app_override(directory=testdirectory)
-
-    # Make a directory where we place the libraries that we would have cloned
-    # if we had use the full waf resolve functionality.
-    #
-    # To avoid relying on network connectivity we simply place the
-    # libraries there and then fake the git clone step.
-    git_dir = testdirectory.mkdir(directory="git_dir")
-
-    qux_dir = mkdir_libqux(directory=git_dir)
-    foo_dir = mkdir_libfoo_json(directory=git_dir)
-    bar_dir = mkdir_libbar(directory=git_dir)
-    baz_dir = mkdir_libbaz(directory=git_dir, qux_dir=qux_dir)
-
-    # Instead of doing an actual Git clone - we fake it and use the paths in
-    # this mapping
-    clone_path = {
-        "acme-corp/foo.git": foo_dir.path(),
-        "acme-corp/bar.git": bar_dir.path(),
-        "acme/baz.git": baz_dir.path(),
-    }
-
-    json_path = os.path.join(app_dir.path(), "clone_path.json")
-
-    with open(json_path, "w") as json_file:
-        json.dump(clone_path, json_file)
-
-    # Try the use checkout
-    app_dir.run(
-        ["python", "waf", "configure", "-v", "--resolve_path", "resolved_dependencies"]
-    )
-    app_dir.run(["python", "waf", "build", "-v"])
-
-    resolve_dir = app_dir.join("resolved_dependencies")
-    assert resolve_dir.contains_dir("foo-*", "1.3.3.7-*")
-    assert resolve_dir.contains_dir("baz-*", "4.0.0-*")
-    assert resolve_dir.contains_dir("bar-*", "someh4sh-*")
-
-
 def test_resolve_only(testdirectory):
     app_dir = mkdir_app(directory=testdirectory)
 
@@ -620,6 +481,7 @@ def test_resolve_only(testdirectory):
     foo_dir = mkdir_libfoo(directory=git_dir)
     bar_dir = mkdir_libbar(directory=git_dir)
     baz_dir = mkdir_libbaz(directory=git_dir, qux_dir=qux_dir)
+    extra_dir = mkdir_libextra(directory=git_dir)
 
     # Instead of doing an actual Git clone - we fake it and use the paths in
     # this mapping
@@ -627,6 +489,7 @@ def test_resolve_only(testdirectory):
         "acme-corp/foo.git": foo_dir.path(),
         "acme-corp/bar.git": bar_dir.path(),
         "acme/baz.git": baz_dir.path(),
+        "acme-corp/extra.git": extra_dir.path(),
     }
 
     json_path = os.path.join(app_dir.path(), "clone_path.json")
@@ -648,7 +511,7 @@ def test_resolve_only(testdirectory):
 
 
 def test_lock_versions_and_then_paths(testdirectory):
-    app_dir = mkdir_app_json(directory=testdirectory)
+    app_dir = mkdir_app(directory=testdirectory)
 
     git_dir = testdirectory.mkdir(directory="git_dir")
 
@@ -656,6 +519,7 @@ def test_lock_versions_and_then_paths(testdirectory):
     foo_dir = mkdir_libfoo(directory=git_dir)
     bar_dir = mkdir_libbar(directory=git_dir)
     baz_dir = mkdir_libbaz(directory=git_dir, qux_dir=qux_dir)
+    extra_dir = mkdir_libextra(directory=git_dir)
 
     # Instead of doing an actual Git clone - we fake it and use the paths in
     # this mapping
@@ -663,10 +527,57 @@ def test_lock_versions_and_then_paths(testdirectory):
         "acme-corp/foo.git": foo_dir.path(),
         "acme-corp/bar.git": bar_dir.path(),
         "acme/baz.git": baz_dir.path(),
+        "acme-corp/extra.git": extra_dir.path(),
     }
 
     with open(os.path.join(app_dir.path(), "clone_path.json"), "w") as json_file:
         json.dump(clone_path, json_file)
+
+    # Check that we cannot run both lock_versions and lock_paths
+    with pytest.raises(RunResultError):
+        r = app_dir.run(
+            [
+                "python",
+                "waf",
+                "configure",
+                "-v",
+                "--lock_versions",
+                "--lock_paths",
+                "--resolve_path",
+                "resolved_dependencies",
+            ],
+        )
+        r.stderr.match("*Incompatible options*")
+
+    # Check that we cannot lock versions or paths while using --skip_internal
+    with pytest.raises(RunResultError):
+        r = app_dir.run(
+            [
+                "python",
+                "waf",
+                "configure",
+                "-v",
+                "--skip_internal",
+                "--lock_versions",
+                "--resolve_path",
+                "resolved_dependencies",
+            ],
+        )
+        r.stderr.match("*Incompatible options*")
+    with pytest.raises(RunResultError):
+        r = app_dir.run(
+            [
+                "python",
+                "waf",
+                "configure",
+                "-v",
+                "--skip_internal",
+                "--lock_path",
+                "--resolve_path",
+                "resolved_dependencies",
+            ],
+        )
+        r.stderr.match("*Incompatible options*")
 
     app_dir.run(
         [
@@ -778,3 +689,54 @@ def test_lock_versions_and_then_paths(testdirectory):
 
     assert app_dir.contains_file("lock_version_resolve.json")
     assert app_dir.contains_file("lock_path_resolve.json")
+
+
+def test_optional(testdirectory):
+    app_dir = mkdir_app(directory=testdirectory)
+
+    git_dir = testdirectory.mkdir(directory="git_dir")
+
+    qux_dir = mkdir_libqux(directory=git_dir)
+    foo_dir = mkdir_libfoo(directory=git_dir)
+    bar_dir = mkdir_libbar(directory=git_dir)
+    baz_dir = mkdir_libbaz(directory=git_dir, qux_dir=qux_dir)
+    extra_dir = mkdir_libextra(directory=git_dir)
+
+    # Instead of doing an actual Git clone - we fake it and use the paths in
+    # this mapping
+    clone_path = {
+        "acme-corp/foo.git": foo_dir.path(),
+        "acme-corp/bar.git": bar_dir.path(),
+        "acme/baz.git": baz_dir.path(),
+        "acme-corp/extra.git": extra_dir.path(),
+    }
+
+    with open(os.path.join(app_dir.path(), "clone_path.json"), "w") as json_file:
+        json.dump(clone_path, json_file)
+
+    r = app_dir.run(
+        [
+            "python",
+            "waf",
+            "configure",
+            "-v",
+            "--resolve_path",
+            "resolved_dependencies",
+        ],
+    )
+
+    assert not r.stdout.match('*Resolve "extra" (git semver)*')
+
+    r = app_dir.run(
+        [
+            "python",
+            "waf",
+            "configure",
+            "-v",
+            "--some_option",
+            "--resolve_path",
+            "resolved_dependencies",
+        ],
+    )
+
+    assert r.stdout.match('*Resolve "extra" (git semver)*')